+++ src/components/filemanager-table/o-filemanager-table.component.ts
@@ -1,23 +1,20 @@
-import { AfterViewInit, Component, forwardRef, Inject, Injector, NgModule, OnDestroy, OnInit, Optional, ViewChild, ViewEncapsulation, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { HttpEventType } from '@angular/common/http';
-
-import { Subscription } from 'rxjs/Subscription';
+import { AfterViewInit, Component, CUSTOM_ELEMENTS_SCHEMA, forwardRef, Inject, Injector, NgModule, OnDestroy, OnInit, Optional, ViewChild, ViewEncapsulation } from '@angular/core';
 import { MatDialog, MatDialogConfig } from '@angular/material';
-
 import { DialogService, InputConverter, OFormComponent, OntimizeWebModule, OSharedModule, OTranslateService } from 'ontimize-web-ngx';
-import { OTableColumnRendererFileTypeComponent } from './renderers/filetype/o-table-column-renderer-filetype.component';
-import { OTableColumnRendererFileSizeComponent } from './renderers/filesize/o-table-column-renderer-filesize.component';
+import { Subscription } from 'rxjs';
 
-import { FileManagerStateService } from '../../services/filemanager-state.service';
-import { OFileManagerTranslateModule, OFileManagerTranslatePipe } from '../../core';
-import { OTableExtendedModule, OTableExtendedComponent } from './table-extended/o-table-extended.component';
-import { ChangeNameDialogComponent, ChangeNameDialogData } from './table-extended/dialog/changename/change-name-dialog.component';
-import { OFileInputExtendedModule, OFileInputExtendedComponent } from '../file-input/o-file-input-extended.component';
-import { FileClass } from '../../core/file.class';
 import { DomService } from '../../services/dom.service';
-import { UploadProgressComponent } from '../status/upload/upload-progress.component';
+import { FileManagerStateService } from '../../services/filemanager-state.service';
+import { FileClass, OFileManagerTranslateModule, OFileManagerTranslatePipe } from '../../util';
+import { OFileInputExtendedComponent, OFileInputExtendedModule } from '../file-input/o-file-input-extended.component';
 import { DownloadProgressComponent } from '../status/download/download-progress.component';
+import { UploadProgressComponent } from '../status/upload/upload-progress.component';
+import { OTableColumnRendererFileSizeComponent } from './renderers/filesize/o-table-column-renderer-filesize.component';
+import { OTableColumnRendererFileTypeComponent } from './renderers/filetype/o-table-column-renderer-filetype.component';
+import { ChangeNameDialogComponent, ChangeNameDialogData } from './table-extended/dialog/changename/change-name-dialog.component';
+import { OTableExtendedComponent, OTableExtendedModule } from './table-extended/o-table-extended.component';
 
 export const DEFAULT_INPUTS_O_FILEMANAGER_TABLE = [
   'workspaceKey: workspace-key',
@@ -27,7 +24,8 @@ export const DEFAULT_INPUTS_O_FILEMANAGER_TABLE = [
   'autoHideTimeout : auto-hide-timeout',
   'serviceType : service-type',
   'newFolderButton: new-folder-button',
-  'selectAllCheckbox: select-all-checkbox'
+  'selectAllCheckbox: select-all-checkbox',
+  'enabled'
 ];
 
 export const DEFAULT_OUTPUTS_O_FILEMANAGER_TABLE = [
@@ -42,7 +40,8 @@ export const DEFAULT_OUTPUTS_O_FILEMANAGER_TABLE = [
   outputs: DEFAULT_OUTPUTS_O_FILEMANAGER_TABLE,
   encapsulation: ViewEncapsulation.None,
   host: {
-    '[class.o-filemanager-table]': 'true'
+    '[class.o-filemanager-table]': 'true',
+    '[class.o-filemanager-table-disabled]': '!enabled'
   },
   providers: [{
     provide: FileManagerStateService,
@@ -64,6 +63,8 @@ export class OFileManagerTableComponent implements OnInit, OnDestroy, AfterViewI
   @InputConverter()
   newFolderButton: boolean = false;
   selectAllCheckbox: string;
+  @InputConverter()
+  public enabled: boolean = true;
 
   queryMethod: string = 'queryFiles';
   deleteMethod: string = 'deleteFiles';
@@ -134,7 +135,7 @@ export class OFileManagerTableComponent implements OnInit, OnDestroy, AfterViewI
     this.translateTable();
 
     if (this.oFileInput) {
-      this.fileChangeSubscription = this.oFileInput.onChange.subscribe(() => {
+      this.fileChangeSubscription = this.oFileInput.onValueChange.subscribe(() => {
         this.showUploaderStatus = true;
         this.doReloadQuery = true;
         let filter = this.stateService.getCurrentQueryFilter();
@@ -336,7 +337,10 @@ export class OFileManagerTableComponent implements OnInit, OnDestroy, AfterViewI
     if (val && this.uploadProgressComponentRef) {
       const instance: UploadProgressComponent = this.uploadProgressComponentRef.instance;
       let files = this.oFileInput.uploader.files.filter(item => !item.isUploaded && !item.isCancel);
+
+
       instance.uploaderFiles = createComp ? files : instance.uploaderFiles.concat(files);
+
       let title = this.translatePipe.transform('MESSAGES.UPLOADING_SINGLE_FILE');
       if (instance.uploaderFiles.length > 1) {
         title = this.translatePipe.transform('MESSAGES.UPLOADING_MULTIPLE_FILE');