+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## 4.0.0 (2019-06-21)
+* **Updated**:  ontimize-web-ngx@latest
+
+## 4.0.0-rc.2
+### PEER-DEPENDENCY UPDATES ###
+* **Updated**:  ontimize-web-ngx@latest
+
+### Features
+* **o-filemanager-table**: new `enabled` attribute ([b87468b](https://github.com/OntimizeWeb/ontimize-web-ngx-filemanager/commit/b87468b)).
+
+## 4.0.0-rc.1 (2019-01-18)
+### Bug Fixes
+* **ontimize-web-ngx**: fixing compatibility bugs for updating to '*ontimize-web-ngx@4.0.0-rc.1*'
+
+## 4.0.0-rc.0 (2018-11-19)
+### Bug Fixes
+* **ontimize-web-ngx**: fixing compatibility bugs for updating to '*ontimize-web-ngx@4.0.0-rc.0*'
+
 ## 3.2.0 (2019-01-16)
 ### BREAKING CHANGES
 * **ontimize-web-ngx**: fixing compatibility bugs for updating to '*ontimize-web-ngx@3.2.0*'
@@ -42,7 +60,7 @@
 
 
 ### Features
-* **OFileManagerModule**: updating library for usign '*ontimize-web-ngx@3.0.0-rc.0*', including update to Angular 5 and Angular Material 5 ([6101eb2](https://github.com/OntimizeWeb/ontimize-web-ngx-charts/commit/6101eb2)) ([cc0d5ca](https://github.com/OntimizeWeb/ontimize-web-ngx-charts/commit/cc0d5ca)) ([c1183f8](https://github.com/OntimizeWeb/ontimize-web-ngx-charts/commit/c1183f8)) ([695e9bf](https://github.com/OntimizeWeb/ontimize-web-ngx-charts/commit/695e9bf)) ([8a9ad78](https://github.com/OntimizeWeb/ontimize-web-ngx-charts/commit/8a9ad78))
+* **OFileManagerModule**: updating library for usign '*ontimize-web-ngx@3.0.0-rc.0*', including update to Angular 5 and Angular Material 5 ([6101eb2](https://github.com/OntimizeWeb/ontimize-web-ngx-filemanager/commit/6101eb2)) ([cc0d5ca](https://github.com/OntimizeWeb/ontimize-web-ngx-filemanager/commit/cc0d5ca)) ([c1183f8](https://github.com/OntimizeWeb/ontimize-web-ngx-filemanager/commit/c1183f8)) ([695e9bf](https://github.com/OntimizeWeb/ontimize-web-ngx-filemanager/commit/695e9bf)) ([8a9ad78](https://github.com/OntimizeWeb/ontimize-web-ngx-filemanager/commit/8a9ad78))
 
 ## 1.0.0